Updated AI to use non-dynamax HP (Leech Seed, Nightmare, Trap Damage, Hail, Sandstorm

+++ include/battle_dynamax.h
@@ -62,7 +62,7 @@ bool32 IsGigantamaxed(u16 battlerId);
 void ApplyDynamaxHPMultiplier(u32 battler, struct Pokemon* mon);
 void PrepareBattlerForDynamax(u16 battlerId);
 u16 GetNonDynamaxHP(u16 battlerId);
-u16 GetNonDynamaxMaxHP(u16 battlerId);
+u16 GetNonDynamaxMaxHP(u32 battlerId);
 void UndoDynamax(u16 battlerId);
 bool32 IsMoveBlockedByMaxGuard(u16 move);
 bool32 IsMoveBlockedByDynamax(u16 move);
